+++ app/controllers/support/cases/documents_controller.rb
@@ -1,0 +1,7 @@
+module Support
+  class Cases::DocumentsController < Cases::ApplicationController
+    def show
+      @document = @current_case.documents.for_rendering.find(params[:id])
+    end
+  end
+end