+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - add service to create SupportCases from SupportEnquiries
 - add ability for agent to resolve a case with notes
 - add ability for agent to send a non templated email on a case
+- add ability for agent to view specification
 - privacy notice page
 - accessibility page